job_agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
playwright>=1.47.0
python-jobspy>=1.1.80
httpx>=0.27.2
pydantic>=2.9.2
python-multipart>=0.0.12
aiofiles>=24.1.0
browser-cookie3>=0.19.1
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.149.0
